fix: move getStaticProps to home page
